functions/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.11.1
chess==1.9.4
pytest==7.4.0